+++ Propiedades.cpp
@@ -10,19 +10,19 @@ int main(){
 	// Recopilación de información:
 	ifstream fin;
 	fin.open("Celdasx_malla.dat");
-		float valor;
-		vector<float> xcr;
+		double valor;
+		vector<double> xcr;
 		while (fin >> valor){
 			xcr.push_back(valor);
 		}
 	fin.close();
 	// Propiedades del material:
-	vector<float> kappa(xcr.size());	// [W/(m*s)]
+	vector<double> kappa(xcr.size());	// [W/(m*s)]
 	for (int i = 0; i < kappa.size(); i++){
 		kappa[i] = 2 * (2.5 - xcr[i] / xcr[xcr.size() - 1]);
 	}
 	// Término fuente:
-	float B = 100;
+	double B = 100;
 	// Escritura de la información:
 	ofstream fout;
 	fout.open("Propiedades.dat");
